feat: standardize voice dictation across app\n\n- Extract robust recording + transcription logic into useVoiceDictation hook\n- Add reusable VoiceDictationButton for mic interactions\n- Add VoiceDictationCard used by Playground (replaces SpeechToTextCard internals)\n- Refactor NewTaskInput to use the shared VoiceDictationButton\n\nThis unifies behavior, fixes inconsistencies (Safari mic stop, MIME types), and\nensures both Playground and Issues pages use the same implementation.
